requirements.txt
Pillow>=6.0.0
textblob>=0.15.3
nltk>=3.4
gensim>=3.8.1
numpy>=1.16.2
pandas>=0.24.2
